requirements.txt
# Python dependencies for the app.
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
reportlab==4.2.5
pypdf>=4.0.0
oracledb==2.3.0
python-dotenv==1.0.1
